requirements.txt
asgiref==3.9.1
colorama==0.4.6
Django==5.0.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-js-asset==3.1.2
django-mptt==0.16.0
django-taggit==5.0.1
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==37.4.2
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.8.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
